passs mark in assignment

+++ LMS_CMS_Backend/LMS_CMS_BL/DTO/HR/VacationEmployeeGetDTO.cs
@@ -19,7 +19,7 @@ namespace LMS_CMS_BL.DTO.HR
         public DateOnly DateFrom { get; set; }
         public DateOnly? DateTo { get; set; }
         public int Balance { get; set; }
-        public decimal used { get; set; }
+        public double used { get; set; }
         public long EmployeeID { get; set; }
         public string EmployeeEnName { get; set; }
         public string EmployeeArName { get; set; }
